+++ src/views/InputView.js
@@ -1,5 +1,5 @@
 const { Console } = require('@woowacourse/mission-utils');
-const MESSAGE = require('./utils/Message');
+const MESSAGE = require('../utils/Message');
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
  */
